move client methods into client.go
